⚡️ ::  memberService 디렉터리 수정 - 소연
⚡️ ::  memberService 디렉터리 수정 - 소연
